+++ CHANGELOG.md
@@ -26,11 +26,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fetching resources or operations with nil team returns personal account
 - Operations not assigning team and project id
-- `view` command to show project label instead of app name
-- `update` command to accept project label instead of app name
 - `export` command to accept project label instead of app name
+- `export` command to list resources without projects
 - `list` command to accept project flag instead of app name
 - `list` now adheres to team context for operations
+- `update` command to accept project label instead of app name
+- `view` command to show project label instead of app name
 - `view` resource still in provision
 
 ### Removed